@@ -11,38 +11,37 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.1",
+    "@angular/animations": "^8.0.0",
     "@angular/cdk": "^7.3.6",
-    "@angular/common": "^7.2.11",
-    "@angular/compiler": "^7.2.11",
-    "@angular/core": "^7.2.11",
-    "@angular/fire": "^5.1.2",
-    "@angular/forms": "^7.2.11",
-    "@angular/http": "^7.2.11",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/fire": "^5.2.1",
+    "@angular/forms": "^8.0.0",
     "@angular/material": "^7.3.3",
-    "@angular/platform-browser": "^7.2.11",
-    "@angular/platform-browser-dynamic": "^7.2.11",
-    "@angular/router": "^7.2.11",
-    "aframe": "^0.9.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
+    "aframe": "^0.9.2",
     "core-js": "^2.5.4",
-    "firebase": "^5.9.2",
-    "rxjs": "~6.3.3",
+    "firebase": "^6.1.0",
+    "rxjs": "~6.5.2",
     "save": "^2.3.3",
     "three": "^0.103.0",
     "three-full": "^11.3.3",
     "tslib": "^1.9.0",
     "update": "^0.7.4",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.4",
-    "@angular/cli": "^7.3.7",
-    "@angular/compiler-cli": "^7.2.11",
-    "@angular/language-service": "^7.2.11",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "^8.0.1",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@types/jasmine": "^3.3.9",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^10.12.30",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.1.4",
@@ -53,6 +52,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.12.1",
-    "typescript": "~3.2.4"
+    "typescript": "^3.4.5"
   }
 }
